Clarify info about IntelliJ-Rust plugin
1) Clarify that plugin is officially supported by JetBrains. This has
not technically happened yet, but we are on the way!
https://github.com/intellij-rust/intellij-rust/issues/1385

2) Describe differences in implementation, rather then differences in
user-visible features. While the latter is surely more important, it
is quite subjective and difficult to keep up to date while both
projects are improving.

3) Remove the note about "heavyweight" IDE :) While IDEA is not
exactly a feather-weight, it's a problem and not a selling point. The
distinguishable feature is that you have complete language model in
your editor, so you can operate with high-level structures like
methods and expressions, rather then characters and words. But this is
also highly subjective :)